improve oneshot style
